fix(review): 代码审阅报告 P0~P2 修复与优化
修复错误：
- compile_pipeline: struct_fields 偏移量从 i*4 硬编码改为按 type_size 累加计算
- compile_pipeline: write_string_to_vm_memory 边界检查修复 null 终止符 off-by-one
- parser: 移除 is_type_token 中错误的 LongLiteral 字面量检测
- parser_unit_test: 修复 assert!(!errors.is_empty() || true) 冗余布尔逻辑
- host_funcs: 统一 host_printf_1/2 的 %f 处理为 f64::from_bits
- flutter_bridge: 修复 get_variables() 实际从 VM 获取变量快照

代码优化：
- lexer: make_token column 计算从 O(n) chars().count() 改为 O(1) text.len()
- bytecode_gen: 字符串字面量内存限制从 0x5000 提升到 0x8000
- algorithm_detector: 增加 MAX_WALK_DEPTH=512 递归深度保护
- error_catalog: lookup_error_info 从 500+ 行 match 改为 LazyLock<HashMap>
- session: 提取 MemoryState::allocate_raw / merge_free_list 公共方法
- host_funcs/vfs: 消除 malloc 与 merge_free_list 的重复代码
- 测试: 移除 5 处 unnecessary_cast，range 检查改为 contains()

验证: cargo test 230/230 通过, cargo clippy --tests 0 警告

+++ native/tests/bytecode_gen_unit_test.rs
@@ -27,7 +27,7 @@ fn test_bytecode_gen_return_const() {
     let _main_idx = output.func_index["main"];
     let main_meta = &output.func_table["main"];
     // Find PushConst 42 and Return in the bytecode around main function
-    let start_ip = main_meta.ip as usize;
+    let start_ip = main_meta.ip;
     let mut found_push = false;
     let mut found_return = false;
     for i in start_ip..output.code.len() {
@@ -54,7 +54,7 @@ fn test_bytecode_gen_local_var() {
 fn test_bytecode_gen_binary_op() {
     let output = generate("int main() { return 1 + 2; }");
     let main_meta = &output.func_table["main"];
-    let start_ip = main_meta.ip as usize;
+    let start_ip = main_meta.ip;
     let mut found_add = false;
     for i in start_ip..output.code.len() {
         if output.code[i].op == OpCode::Add {
@@ -69,7 +69,7 @@ fn test_bytecode_gen_binary_op() {
 fn test_bytecode_gen_if_statement() {
     let output = generate("int main() { if (1) { return 1; } return 0; }");
     let main_meta = &output.func_table["main"];
-    let start_ip = main_meta.ip as usize;
+    let start_ip = main_meta.ip;
     let mut found_jump_if_zero = false;
     let mut found_jump = false;
     for i in start_ip..output.code.len() {
@@ -87,7 +87,7 @@ fn test_bytecode_gen_if_statement() {
 fn test_bytecode_gen_while_loop() {
     let output = generate("int main() { while (0) { } return 0; }");
     let main_meta = &output.func_table["main"];
-    let start_ip = main_meta.ip as usize;
+    let start_ip = main_meta.ip;
     let mut found_jump_if_zero = false;
     for i in start_ip..output.code.len() {
         if output.code[i].op == OpCode::JumpIfZero {
@@ -104,7 +104,7 @@ fn test_bytecode_gen_function_call() {
     assert!(output.func_index.contains_key("add"));
     assert!(output.func_index.contains_key("main"));
     let main_meta = &output.func_table["main"];
-    let start_ip = main_meta.ip as usize;
+    let start_ip = main_meta.ip;
     let mut found_call = false;
     for i in start_ip..output.code.len() {
         if output.code[i].op == OpCode::Call {
